Language.PlutusCore.Interpreter.CekMachine

Description

The CEK machine. Rules are the same as for the CK machine from Language.PlutusCore.Evaluation.CkMachine, except we do not use substitution and use environments instead. The CEK machine relies on variables having non-equal Uniques whenever they have non-equal string names. I.e. Uniques are used instead of string names. This is for efficiency reasons. The type checker pass is a prerequisite. Feeding ill-typed terms to the CEK machine will likely result in a MachineException. The CEK machine generates booleans along the way which might contain globally non-unique Uniques. This is not a problem as the CEK machines handles name capture by design. Dynamic extensions to the set of built-ins are allowed. In case an unknown dynamic built-in is encountered, an UnknownDynamicBuiltinNameError is returned (wrapped in OtherMachineError).

Documentation

type CekMachineException = MachineException UnknownDynamicBuiltinNameError

The CEK machine-specific MachineException.

data EvaluationResultF a

The parameterized type of results various evaluation engines return.

Constructors

EvaluationSuccess a

EvaluationFailure

type EvaluationResult = EvaluationResultF (Value TyName Name ())

The type of results various evaluation engines return.

evaluateCekCatch :: DynamicBuiltinNameMeanings -> Plain Term -> Either CekMachineException EvaluationResult

Evaluate a term using the CEK machine.

evaluateCek :: DynamicBuiltinNameMeanings -> Term TyName Name () -> EvaluationResult

Evaluate a term using the CEK machine. May throw a CekMachineException.

readDynamicBuiltinCek :: KnownDynamicBuiltinType dyn => Term TyName Name () -> Either CekMachineException (Maybe dyn)

runCek :: DynamicBuiltinNameMeanings -> Program TyName Name () -> EvaluationResult

Run a program using the CEK machine. May throw a CekMachineException. Calls evaluateCek under the hood.