Language.PlutusCore.Interpreter.LMachine

Description

The L machine A lazy machine based on the L machine of Friedman et al. [Improving the Lazy Krivine Machine] For more details see the document in plutusdocsfomega/lazy-machine The code here's closely based on the CEK machine implementation.

Documentation

data EvaluationResultF a

The parameterized type of results various evaluation engines return.

Constructors

EvaluationSuccess a

EvaluationFailure

type EvaluationResult = EvaluationResultF (Value TyName Name ())

The type of results various evaluation engines return.

evaluateL :: Term TyName Name () -> EvaluationResult

Evaluate a term using the L machine. May throw an MachineException.

runL :: DynamicBuiltinNameMeanings -> Program TyName Name () -> EvaluationResult

Run a program using the L machine. May throw a MachineException. We're not using the dynamic names at the moment, but we'll require them eventually.